ラジオボタンとリスト表示の実装をコンポーネントに変更
